Fixed the donut charts distibution metric

+++ PennyPilot.Backend/PennyPilot.Backend.Application/DTOs/DonutChartsDto.cs
@@ -8,9 +8,9 @@ namespace PennyPilot.Backend.Application.DTOs
 {
     public class DonutChartsDto
     {
-        public Dictionary<string, int>? ExpenseCategories { get; set; }        
+        public Dictionary<string, decimal>? ExpenseCategories { get; set; }        
         public Dictionary<string, decimal>? UserExpenses { get; set; }
-        public Dictionary<string, int>? IncomeCategories { get; set; }
+        public Dictionary<string, decimal>? IncomeCategories { get; set; }
         public Dictionary<string, decimal>? IncomeSources { get; set; }
     }
 }
